refactor : 구조개선 > ioc 컨테이너를 외부로 노출하지 않음
