+++ src/main/java/carracing/service/RacingService.java
@@ -1,13 +1,18 @@
 package carracing.service;
 
+import carracing.domain.Record;
 import carracing.domain.Round;
 import carracing.domain.Score;
 import carracing.repository.RoundRepository;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+
+import static carracing.RaceApplication.getIoc;
 
 public class RacingService {
     private static final Random random = new Random();
@@ -31,16 +36,41 @@ public class RacingService {
         return roundRepository.findAll();
     }
 
-    private void roundStart(int participates) {
+    public void roundStart(int participates) {
+        List<Score> collect = IntStream.rangeClosed(1, participates)
+            .boxed()
+            .map(s -> randomScore())
+            .map(Score::new)
+            .collect(Collectors.toList());
         roundRepository.save(new Round(
-            IntStream.rangeClosed(1, participates)
-                .boxed()
-                .map(s -> randomScore())
-                .map(Score::new)
-                .collect(Collectors.toList()))
-        );
+            collect, scoreToRecord(collect)
+        ));
     }
 
+    private List<Record> scoreToRecord(List<Score> scoreList) {
+        AtomicInteger integer = new AtomicInteger(0);
+        List<Record> recordList = new ArrayList<>();
+        for(Score score : scoreList) {
+            recordList.add(new Record(calRecord(integer, score)));
+        }
+        return recordList;
+    }
+
+    private String calRecord(AtomicInteger integer, Score score) {
+        if(isFirstRound() ) {
+            return score.toProgress();
+        }
+        return score.toProgress() + prevScoreProgress(integer.getAndIncrement());
+    }
+
+    private boolean isFirstRound() {
+        return roundRepository.count() == 0;
+    }
+
+    private String prevScoreProgress(int index) {
+        Round byId = roundRepository.findById(roundRepository.count());
+        return byId.getRecords().get(index).getRecord();
+    }
     public List<Round> findAll() {
         return roundRepository.findAll();
     }
